+++ sdk/data/azcosmos/cosmos_global_endpoint_manager_test.go
@@ -6,6 +6,7 @@ package azcosmos
 import (
 	"context"
 	"encoding/json"
+	"fmt"
 	"net/http"
 	"net/url"
 	"strings"
@@ -445,6 +446,51 @@ func TestRequestToUpdateGEMPreservesIncomingContextWithoutCancellation(t *testin
 	if value != "testValue" {
 		t.Fatalf("expected a captured context to contain test=testValue, got test=%v", value)
 	}
+}
+
+func TestAddedAllowTentativeHeaderGEMPolicy(t *testing.T) {
+	type contextKey string
+
+	gemServer, gemClose := mock.NewTLSServer()
+	defer gemClose()
+	gemServer.SetResponse(mock.WithStatusCode(200))
+	serverEndpoint, _ := url.Parse("https://myaccount.documents.azure.com:443/")
+	mocked_response := "{\"_self\":\"\",\"id\":\"my_account\",\"_rid\":\"my_account-westus.sql.cosmos.azure.com\",\"media\":\"//media/\",\"addresses\":\"//addresses/\",\"_dbs\":\"//dbs/\",\"writableLocations\":[{\"name\":\"West US\",\"databaseAccountEndpoint\":\"https://my_account-westus.documents.azure.com:443/\"},{\"name\":\"West US 3\",\"databaseAccountEndpoint\":\"https://my_account-westus3.documents.azure.com:443/\"}],\"readableLocations\":[{\"name\":\"West US\",\"databaseAccountEndpoint\":\"https://my_account-westus.documents.azure.com:443/\"},{\"name\":\"West US 3\",\"databaseAccountEndpoint\":\"https://my_account-westus3.documents.azure.com:443/\"}], \"enableMultipleWriteLocations\":true}"
+
+	gemServer.SetResponse(mock.WithBody([]byte(mocked_response)))
+	mockLc := createLocationCacheForGem(*serverEndpoint, true)
+
+	// The GEM needs it's own pipeline that doesn't have the GEM policy in it to avoid deadlocking.
+	capturePolicy := &requestCollector{}
+	gemPipeline := azruntime.NewPipeline("azcosmosgemtest", "v1.0.0", azruntime.PipelineOptions{}, &policy.ClientOptions{Transport: gemServer, PerCallPolicies: []policy.Policy{capturePolicy}})
+	mockGem := &globalEndpointManager{
+		clientEndpoint:      gemServer.URL(),
+		pipeline:            gemPipeline,
+		preferredLocations:  []string{"Central US"},
+		locationCache:       mockLc,
+		refreshTimeInterval: 5 * time.Minute,
+	}
+
+	gemPolicy := &globalEndpointManagerPolicy{
+		gem: mockGem,
+	}
+
+	// For the "main" pipeline under test, we can insert the GEM policy, which will cause GEM updates to run (through the GEM pipeline).
+	testPipeline := azruntime.NewPipeline("azcosmosgemtest", "v1.0.0", azruntime.PipelineOptions{}, &policy.ClientOptions{Transport: gemServer, PerCallPolicies: []policy.Policy{gemPolicy}})
+
+	ctx, cancel := context.WithCancel(context.WithValue(context.Background(), contextKey("test"), "testValue"))
+	defer cancel()
+
+	// Issue a test request
+	req, err := azruntime.NewRequest(ctx, http.MethodGet, gemServer.URL())
+	if err != nil {
+		t.Fatalf("failed to create request: %v", err)
+	}
+	resp, _ := testPipeline.Do(req)
+	if resp.Request.Header.Get(cosmosHeaderAllowTentativeWrites) == "" {
+		t.Fatalf("expected %s header to be set", cosmosHeaderAllowTentativeWrites)
+	}
+	fmt.Println(resp)
 }
 
 func createLocationCacheForGem(defaultEndpoint url.URL, isMultiMaster bool) *locationCache {